+++ src/structures/index.ts
@@ -1,7 +1,6 @@
 export * from './BaseStructure';
 export * from './Channel';
 export * from './ClientUser';
-export * from './Embed';
 export * from './Guild';
 export * from './GuildMember';
 export * from './Message';
